requirements.txt
flask==3.0.3
torch==2.2.2
torchvision==0.17.2
transformers==4.41.2
pillow==10.3.0
numpy==1.26.4
opencv-python==4.9.0.80
tqdm==4.66.4
requests==2.31.0
scipy==1.13.1